The last version of ibus 1.5.3 breaks my system.
I use Super-\\ for activating IME, it can't be used with my scheme of
using sawfish synthesize-event to change it to C-| anymore.

I'm currently using xdotool to work around this. Seems it is workable
if I use Super-F12 as the hotkey in ibus, then I map Super-\\ in
sawfish to run xdotool key super+F12, as long as I sleep .1 before
xdotool (so that when xdotool runs, the \\ key is already
released. The super key does not matter whether it is still pressed or
released, since it is needed by ibus anyway).

Signed-off-by: Bao Haojun <baohaojun@gmail.com>