Handle language name symbols in the blocks list
When resolving the fenced code block at point, we were checking
whether the language identifier was a bound function, which I think
was a previous strategy of specifying actual mode function names. That
strategy was later changed to "language identifier symbol," so this
updates the function to treat that symbol as such.

Fixes #2