Fix a critical bug for non-system-config setup.
Should not use ew (edit-wait), as it is from system-config.

Also, added a erwo sub-command, for editing a region as org-mode.

This is useful for e.g., when I'm writing comment in a shell script, I
can select the current comment, and strip the leading '#', and edit it
as org-mode. See https://github.com/aaronbieber/fence-edit.el.